# backend/swagger/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Tarot API Test",
    "description": "오늘도 화이팅",
    "version": "1.0.0"
  },
  "host": "43.202.208.226:3001",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "JWT Access Token 입력 : 'Bearer {token}'"
    }
  },
  "paths": {
...
    "/mypage/detail": {
      "get": {
        "tags": [
          "MyPage"
        ],
        "summary": "결과 리스트에서 선택한 결과 상세 조회",
        "description": "결과 리스트 중 선택한 결과(Poll_id)를 통해 해당 결과를 상세조회한다.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "poll_id",
            "in": "query",
            "description": "poll_id",
            "required": true,
            "example": "12",
            "value": "12",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "총합 결과와 카드별 해석 및 링크 데이터 성공적으로 조회됨.",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "question": {
                        "type": "string",
                        "example": "테스트 질문"
                      },
                      "explanation": {
                        "type": "string",
                        "example": "테스트 종합 해석"
                      },
                      "luck": {
                        "type": "string",
                        "example": "테스트 럭 리스트"
                      },
                      "master_name": {
                        "type": "string",
                        "example": "테스트 마스터"
                      },
                      "created_at": {
                        "type": "string",
                        "example": "2024-01-11"
                      }
                    }
                  }
                },
                "card": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "image_url": {
                        "type": "string",
                        "example": "테스트url"
                      },
                      "explanation": {
                        "type": "string",
                        "example": "카드해석"
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "내부 서버 오류로 인한 DB 쿼리 실패.",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "DB 쿼리 오류"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/mypage/delete": {
      "delete": {
        "tags": [
          "MyPage"
        ],
        "summary": "결과 리스트에서 선택한 결과 및 카드 삭제",
        "description": "결과 리스트 중 선택한 결과(Poll_id) 및 해당 결과에 대한 카드를 삭제한다.",
        "parameters": [
          {
            "name": "poll_id",
            "in": "query",
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  }
}